+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG  
 
+## [ 1.0.1 ]
+- Add + to value displayed on popup
+- Minor change to CSS for message to center text & bold CRIT messages
+
 ## [ 1.0.0]
 - Initial Commit